backend/foodgram/requirements.txt
asgiref==3.7.2
Django==3.2.20
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
djoser==2.1.0
Pillow==9.5.0
django-filter==23.2
django-extra-fields==3.0.2
psycopg2-binary==2.9.7
python-dotenv==0.21.1
gunicorn==21.2.0